requirements.txt
pycryptodome==3.10.1
pyperclip==1.8.2
UnityPy==1.7.8
